Times serialization/deserialization operations
I've updated the ReadStreams and WriteStreams
methods to time how long the respective operations
take.

For reading, I'm checking elapsed time after the
execution of the loop but before the stream is
closed, because I figure the work I care about
there is already done.

For writing, I'm checking elapsed time after the
stream is closed, because I'd consider time to
flush to disk important for serialization. This
may be less important for either memory backed
stream (effectively eliminating the flush time)
or a network backed stream (making 99.9% of
the operation time *be* the network), both
things worth checking.

For both operations, I clearly need to learn
how to properly use the new JSON API. Questions
for myself include,

1) Does myself as a user of the new System.Text
JSON API need to familiarize myself with
Span<T> and Memory<T> to effectively use it, or
is just pointing the results of a StreamReader
sufficient to gain it's performance benefits?
2) What about more complex data, like a series
of more complex object graphs (think `contacts`,
`job history`, `relationships` for people, in
this example).